+++ logicnode_definitions/__init__.py
@@ -1,0 +1,5 @@
+# Import all nodes
+from os.path import dirname, basename, isfile
+import glob
+modules = glob.glob(dirname(__file__)+"/*.py")
+__all__ = [basename(f)[:-3] for f in modules if isfile(f)]
